Saldo page works!!!

+++ BankCode/BankJava/src/mypackage/JPanel05.java
@@ -154,8 +154,8 @@ public class JPanel05 extends javax.swing.JPanel {
         // (A) Terug naar begin
     }//GEN-LAST:event_jButton1ActionPerformed
 
-    public void changeBalanceLabel(int balance){
-        jLabel3.setText("$ " + Integer.toString(balance));
+    public void changeBalanceLabel(float balance){
+        jLabel3.setText("$ " + Float.toString(balance));
     }
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
